+++ src/fx/consume_tick.rs
@@ -1,7 +1,7 @@
 use ratatui::buffer::Buffer;
 use ratatui::layout::Rect;
 
-use crate::{CellIterator, Duration};
+use crate::Duration;
 use crate::CellFilter;
 use crate::shader::Shader;
 
@@ -25,8 +25,6 @@ impl Shader for ConsumeTick {
         self.has_consumed_tick = true;
         None
     }
-
-    fn execute(&mut self, _alpha: f32, _area: Rect, _cell_iter: CellIterator) {}
 
     fn done(&self) -> bool { self.has_consumed_tick }