+++ src/fx/dissolve.rs
@@ -1,10 +1,10 @@
+use ratatui::buffer::Buffer;
 use ratatui::layout::Rect;
 
 use crate::effect_timer::EffectTimer;
 use crate::shader::Shader;
 use crate::simple_rng::SimpleRng;
 use crate::CellFilter;
-use crate::CellIterator;
 
 #[derive(Clone)]
 pub struct Dissolve {
@@ -32,7 +32,8 @@ impl Shader for Dissolve {
         if self.timer.is_reversed() { "coalesce" } else { "dissolve" }
     }
 
-    fn execute(&mut self, alpha: f32, _area: Rect, cell_iter: CellIterator) {
+    fn execute(&mut self, alpha: f32, area: Rect, buf: &mut Buffer) {
+        let cell_iter = self.cell_iter(buf, area);
         let mut lcg = self.lcg;
         cell_iter
             .filter(|_| alpha > lcg.gen_f32())