+++ src/fx/ping_pong.rs
@@ -1,7 +1,7 @@
 use ratatui::buffer::Buffer;
 use ratatui::layout::Rect;
 
-use crate::{CellFilter, CellIterator, Duration, Effect, EffectTimer, Shader};
+use crate::{CellFilter, Duration, Effect, EffectTimer, Shader};
 use crate::widget::EffectSpan;
 
 #[derive(Clone)]
@@ -42,10 +42,6 @@ impl Shader for PingPong {
         } else {
             overflow
         }
-    }
-
-    fn execute(&mut self, _alpha: f32, _area: Rect, _cell_iter: CellIterator) {
-        // nothing to do
     }
 
     fn done(&self) -> bool {